refactor: move svg files
